refactor: move chatsystem to its own package
